+++ src/Config_Definitions.h
@@ -26,7 +26,7 @@ using namespace std;
 #endif
 
 // Software version
-const string VERSION = "1.9.1.7";
+const string VERSION = "1.9.1.8";
 
 // Battery indicator
 // Note: This battery can be for the Ignitron controller or the Spark amp.